requirements.txt
# python3.11 | railway
flask==2.3.3
gunicorn==21.2.0
python-decouple==3.8
python-dotenv==1.0.0
supabase==1.0.3
